Fixes #37 - save falsy checkbox values
The code in `_initServices()` uses `undefined` as expected -- to indicate that
a field doesn't exist. OTOH, when `onUpdateService` saves unchecked checkboxes,
it deletes their value from settings, which also results in `undefined`.

This patch changes that code such that a `newValue` of `undefined` does indeed
delete the value of the setting, but a falsy `newValue` is still written to
the settings.